Refactor the way Shop Page is built
