+++ src/components/CollectionsOverview/CollectionsOverview.styles.scss
@@ -1,0 +1,4 @@
+.collections-overview {
+  display: flex;
+  flex-direction: column;
+}
