+++ src/pages/Shop/Shop.page.jsx
@@ -1,20 +1,11 @@
-import { useSelector } from "react-redux";
-import { createStructuredSelector } from "reselect";
-import { selectShopItems } from "../../redux/reducers/shop/shop.selectors";
-
 import "./Shop.styles.scss";
 
-import { PreviewCollection } from "../../components/components";
-const collectionSelector = createStructuredSelector({
-  collections: selectShopItems,
-});
+import { CollectionsOverview } from "../../components/components";
+
 const ShopPage = () => {
-  const { collections } = useSelector(collectionSelector);
   return (
     <div className="shop-page">
-      {collections.map(({ id, ...otherCollectionProps }) => (
-        <PreviewCollection key={id} {...otherCollectionProps} />
-      ))}
+      <CollectionsOverview />
     </div>
   );
 };
